+++ monster_service/src/main/java/dnd/monster_service/game/logic/encounter/encounter_difficulty/EncounterDifficultyMap.java
@@ -8,7 +8,7 @@ public interface EncounterDifficultyMap {
 
     Double getMultiplier(Integer number);
 
-    Integer getXp(Double cr);
+    Integer getXp(float cr);
 
     /**
         Returns cr closest to xp
