+++ monster_service/src/main/java/dnd/monster_service/persistance/model/creature/monster/Monster.java
@@ -32,7 +32,7 @@ public class Monster extends Creature{
     @Id
     @GeneratedValue
     private long id;
-    private Double cr;
+    private Float cr;
     private String monsterName;
     private Short passivePerception;
     private String imageUrl;
